+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/BlocksEditor.tsx
@@ -18,6 +18,7 @@ import { paragraphBlocks } from './Blocks/Paragraph';
 import { quoteBlocks } from './Blocks/Quote';
 import { BlocksContent } from './BlocksContent';
 import { BlocksToolbar } from './BlocksToolbar';
+import { withImages } from './plugins/withImages';
 import { withLinks } from './plugins/withLinks';
 import { withStrapiSchema } from './plugins/withStrapiSchema';
 
@@ -135,22 +136,6 @@ function useResetKey(value?: Attribute.BlocksValue): {
 
   return { key, incrementSlateUpdatesCount: () => (slateUpdatesCount.current += 1) };
 }
-
-/**
- * Images are void elements. They handle the rendering of their children instead of Slate.
- * See the Slate documentation for more information:
- * - https://docs.slatejs.org/api/nodes/element#void-vs-not-void
- * - https://docs.slatejs.org/api/nodes/element#rendering-void-elements
- */
-const withImages = (editor: Editor) => {
-  const { isVoid } = editor;
-
-  editor.isVoid = (element) => {
-    return element.type === 'image' ? true : isVoid(element);
-  };
-
-  return editor;
-};
 
 const pipe =
   (...fns: ((baseEditor: Editor) => Editor)[]) =>
