+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/plugins/withImages.ts
@@ -1,0 +1,19 @@
+import { type Editor } from 'slate';
+
+/**
+ * Images are void elements. They handle the rendering of their children instead of Slate.
+ * See the Slate documentation for more information:
+ * - https://docs.slatejs.org/api/nodes/element#void-vs-not-void
+ * - https://docs.slatejs.org/api/nodes/element#rendering-void-elements
+ */
+const withImages = (editor: Editor) => {
+  const { isVoid } = editor;
+
+  editor.isVoid = (element) => {
+    return element.type === 'image' ? true : isVoid(element);
+  };
+
+  return editor;
+};
+
+export { withImages };
